+++ src/lib.rs
@@ -52,7 +52,7 @@ const PLAYER_DMG_STAT: f32 = 10.0;
 const ENEMY_SIZE: f32 = 30.0;
 const PLAYER_SIZE: f32 = 30.0;
 const SPELL_SIZE: f32 = 16.0;
-const XP_GAIN_GEM: i32 = 10;
+const XP_GAIN_GEM: f32 = 10.;
 
 /// How quickly should the camera snap to the desired location.
 const CAMERA_DECAY_RATE: f32 = 2.;
